Create a utility library for working with nodejs
There are currently fable bindings to nodejs, but they reflect the OO nature of node libs (as they should).

Instead, we can write a high-level abstraction lib that starts to let us work with node in a more functional way.

Signed-off-by: Joe Grund <joe.grund@intel.com>

Create a utility library for working with nodejs

There are currently fable bindings to nodejs, but they reflect the OO nature of node libs (as they should).

Instead, we can write a high-level abstraction lib that starts to let us work with node in a more functional way.

Signed-off-by: Joe Grund <joe.grund@intel.com>