Fixup node streams to autoconfigure
Streams can be in different modes. We can use generics in the type
system to show what the intended mode is.

However, streams also need to be configured in runtime mode to
properly deal with incoming / outgoing data.

We can use a combination of reflection and stream introspection
to configure the streams properly.

Also update Node bindings to latest.

Signed-off-by: Joe Grund <joe.grund@intel.com>